Added utility for GetDroppedFiles. - Used a function from a older version that converts a IntPtr returned by GetDroppedFiles to a string[]. - Updated examples that use GetDroppedFiles(ref count); to use Utils.MarshalDroppedFiles(ref count); instead.
The length of each string can vary based on how raylib is compiled so I cannot do

[return: MarshalAs(UnmanagedType.LPArray, ArraySubType=UnmanagedType.LPStr, SizeConst=512)]

as the size could be different and it has to be a constant value.

For now this works and I am leaving it in Utils.cs. This can be revisited in the future if it a nicer function is found.